requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pydantic-settings
authlib
httpx
itsdangerous
bcrypt
jinja2
python-multipart
aiofiles
pillow
celery
redis
weasyprint
